+++ client/src/shared/PanelHelper.cs
@@ -209,7 +209,16 @@ namespace OpenGaugeClient
 
         public static bool GetIsPanelVisible(Panel panel, string? vehicleName)
         {
-            return (panel.Vehicle != null && vehicleName != null && Utils.GetIsVehicle(panel.Vehicle, vehicleName)) || panel.Vehicle == null;
+            if (panel.Force == true)
+                return true;
+
+            if (vehicleName == null)
+                return false;
+
+            if (panel.Vehicle != null)
+                return Utils.GetIsVehicle(panel.Vehicle, vehicleName);
+
+            return true;
         }
     }
 }